refactor(golint): put the files inside cmd/foo and add more go code styles
